feat: create cart & not found pages
