+++ package.json
@@ -14,7 +14,8 @@
   "dependencies": {
     "react": "^18.3.1",
     "react-content-loader": "^7.0.2",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.2"
   },
   "devDependencies": {
     "@eslint/js": "^9.9.0",
