+++ src/pages/Cart.tsx
@@ -1,0 +1,3 @@
+export const Cart = () => {
+  return <div>Cart</div>;
+};
