+++ src/pages/NotFound.tsx
@@ -1,0 +1,3 @@
+export const NotFound = () => {
+  return <h1>Ничего не найдено</h1>;
+};
